DotStar support

+++ src/internal/HtmlColor.h
@@ -90,7 +90,7 @@ struct HtmlColor
     // ------------------------------------------------------------------------
     HtmlColor(const RgbColor& color)
     {
-        Color = color.R << 16 | color.G << 8 | color.B;
+        Color = (uint32_t)color.R << 16 | (uint32_t)color.G << 8 | (uint32_t)color.B;
     }
 
     // ------------------------------------------------------------------------
